Added NaaradUtils for global functions+added timeout for the "notify" command.
